requirements.txt
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.6
django-bootstrap3==9.0.0
django-cleanup==1.0.1
django-debug-toolbar==1.5
django-static-precompiler==1.6
django-webpack-loader==0.5.0
djangorestframework==3.7.0
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.25.0
google-cloud-storage==1.2.0
google-resumable-media==0.1.1
googleapis-common-protos==1.5.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
mysqlclient==1.3.12
oauth2client==4.1.2
olefile==0.44
Pillow==4.1.1
protobuf==3.3.0
psycopg2==2.7.3.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
Pygments==2.2.0
pytz==2017.2
PyYAML==3.12
requests==2.18.1
rsa==3.4.2
six==1.10.0
sqlparse==0.2.3
urllib3==1.21.1
whitenoise==3.0
yamjam==0.1.7
